Remember latest message timestamp/id, and fetch only newer messages. The messageID is now stored as a scalar (int32_t).

+++ Shared/MessageDataHandler.m
@@ -66,7 +66,7 @@
 				
 				Message *msg = [[Message alloc] init];
 				msg.timestamp = date;
-				msg.messageID = messageID;
+				msg.messageID = messageID.intValue;
 				msg.topic = topic;
 				msg.content = content;
 				[messageList addObject:msg];
